+++ src/flownet/ahm/_run_ahm.py
@@ -416,14 +416,20 @@ def run_flownet_history_matching(
         field_data.faults if config.model_parameters.fault_mult else None
     )
 
-    concave_hull_bounding_boxes: Optional[np.ndarray] = None
+    concave_hull_list: Optional[List[np.ndarray]] = None
     if config.flownet.data_source.concave_hull:
-        concave_hull_bounding_boxes = field_data.grid_cell_bounding_boxes
+        concave_hull_list = []
+        for layer_id in df_well_connections["LAYER_ID"].unique():
+            concave_hull_list.append(
+                field_data.grid_cell_bounding_boxes(layer_id=layer_id)
+            )
 
     df_entity_connections: pd.DataFrame = create_connections(
-        df_well_connections[["WELL_NAME", "X", "Y", "Z"]].drop_duplicates(keep="first"),
+        df_well_connections[["WELL_NAME", "X", "Y", "Z", "LAYER_ID"]].drop_duplicates(
+            keep="first"
+        ),
         config,
-        concave_hull_bounding_boxes=concave_hull_bounding_boxes,
+        concave_hull_list=concave_hull_list,
     )
 
     network = NetworkModel(